fix enum import issues and add mechanism to retry repos marked as Standby
